+++ modules/components/operators/Proximity.jsx
@@ -29,7 +29,7 @@ export default class Proximity extends PureComponent {
   };
 
   handleChange = (value) => {
-    this.props.setOption('proximity', value);
+    this.props.setOption('proximity', parseInt(value));
   }
 
   render() {