+++ modules/components/widgets/antd/Number.jsx
@@ -5,15 +5,16 @@ import { InputNumber, Col } from 'antd';
 export default class NumberWidget extends PureComponent {
   static propTypes = {
     setValue: PropTypes.func.isRequired,
-    min: PropTypes.number,
-    max: PropTypes.number,
-    step: PropTypes.number,
     placeholder: PropTypes.string,
     config: PropTypes.object.isRequired,
     field: PropTypes.string.isRequired,
     value: PropTypes.number,
     customProps: PropTypes.object,
     fieldDefinition: PropTypes.object,
+    // from fieldSettings:
+    min: PropTypes.number,
+    max: PropTypes.number,
+    step: PropTypes.number,
   };
 
   handleChange = (val) => {
@@ -29,14 +30,9 @@ export default class NumberWidget extends PureComponent {
   };
 
   render() {
-    const {config, placeholder, fieldDefinition, customProps, value,  min, max, step} = this.props;
+    const {config, placeholder, customProps, value,  min, max, step} = this.props;
     const {renderSize} = config.settings;
-    const {fieldSettings} = fieldDefinition || {};
-
     const _value = value != undefined ? value : undefined;
-    const _min = fieldSettings.min != null ? fieldSettings.min : min;
-    const _max = fieldSettings.max != null ? fieldSettings.max : max;
-    const _step = fieldSettings.step != null ? fieldSettings.step : step;
 
     return (
       <Col>
@@ -45,9 +41,9 @@ export default class NumberWidget extends PureComponent {
           size={renderSize}
           ref="num"
           value={_value}
-          min={_min}
-          max={_max}
-          step={_step}
+          min={min}
+          max={max}
+          step={step}
           placeholder={placeholder}
           onChange={this.handleChange}
           {...customProps}